--- correctGameStateTest.dodge.txt
@@ -1,6 +1,0 @@
-16
-12
-0
-false
-.................................o...............o...............o...............o......................x.......................................................................................
-UP,DOWN,LEFT,RIGHT
