+++ src/test/java/testFile/correctGameStateTest.dodge.txt
@@ -1,0 +1,6 @@
+16
+12
+0
+false
+.................................o...............o...............o...............o......................x.......................................................................................
+UP,DOWN,LEFT,RIGHT
